requirements.txt
jupyterlab
numpy
tqdm
ultralytics
python-dotenv
opencv-python